+++ api/code/traffic.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
+	"net"
 	"net/http"
 	"os/exec"
 	"sync"
@@ -13,12 +14,14 @@ import (
 import (
 	"github.com/gorilla/mux"
 	"github.com/influxdata/influxdb-client-go/v2"
+	"github.com/spr-networks/sprbus"
 )
 
 var IFDB influxdb2.Client = nil
 
 type TrafficElement struct {
 	IP      string
+	Domain  string
 	Packets uint64
 	Bytes   uint64
 }
@@ -27,6 +30,8 @@ type IPTrafficElement struct {
 	Interface string
 	Src       string
 	Dst       string
+	SrcDomain string
+	DstDomain string
 	Packets   uint64
 	Bytes     uint64
 	Timeout   uint64
@@ -42,8 +47,17 @@ func parseTrafficElements(elements []interface{}) []TrafficElement {
 			if ok {
 				counter, ok := ele["counter"].(map[string]interface{})
 				if ok {
+					ip := ele["val"].(string)
+					domain := ""
+					DNSCachemtx.RLock()
+					src_domain, exists := DNSCache[ip]
+					if exists {
+						domain = src_domain
+					}
+					DNSCachemtx.RUnlock()
 					traffic = append(traffic,
-						TrafficElement{IP: ele["val"].(string),
+						TrafficElement{IP: ip,
+							Domain:  domain,
 							Bytes:   uint64(counter["bytes"].(float64)),
 							Packets: uint64(counter["packets"].(float64))})
 				}
@@ -74,14 +88,30 @@ func parseIPTrafficElements(elements []interface{}) []IPTrafficElement {
 							src_ip, _ := concat[1].(string)
 							dst_ip, _ := concat[2].(string)
 
+							SrcDomain := ""
+							DstDomain := ""
+							DNSCachemtx.RLock()
+							src_domain, exists := DNSCache[src_ip]
+							if exists {
+								SrcDomain = src_domain
+							}
+							dst_domain, exists := DNSCache[dst_ip]
+							if exists {
+								DstDomain = dst_domain
+							}
+							DNSCachemtx.RUnlock()
+
 							traffic = append(traffic,
 								IPTrafficElement{Interface: ifname,
-									Src:     src_ip,
-									Dst:     dst_ip,
-									Bytes:   uint64(counter["bytes"].(float64)),
-									Packets: uint64(counter["packets"].(float64)),
-									Timeout: uint64(timeout),
-									Expires: uint64(expires)})
+									Src:       src_ip,
+									Dst:       dst_ip,
+									SrcDomain: SrcDomain,
+									DstDomain: DstDomain,
+									Bytes:     uint64(counter["bytes"].(float64)),
+									Packets:   uint64(counter["packets"].(float64)),
+									Timeout:   uint64(timeout),
+									Expires:   uint64(expires)})
+
 						}
 					}
 				}
@@ -341,4 +371,126 @@ func initTraffic(config APIConfig) {
 	if config.InfluxDB.URL != "" && config.InfluxDB.Token != "" {
 		IFDB = influxdb2.NewClient(config.InfluxDB.URL, config.InfluxDB.Token)
 	}
+}
+
+type DNSRecord struct {
+	A           []ARecord
+	FirstAnswer string
+	FirstName   string
+	Local       string
+	Q           []QueryRecord
+	Remote      string
+	Timestamp   string
+	Type        string
+	Time        string
+	Bucket      string
+}
+
+type ARecord struct {
+	A   string
+	Hdr Header
+}
+
+type Header struct {
+	Class    int
+	Name     string
+	Rdlength int
+	Rrtype   int
+	Ttl      int
+}
+
+type QueryRecord struct {
+	Name   string
+	Qclass int
+	Qtype  int
+}
+
+// ip -> domain map
+var DNSCache = map[string]string{}
+var DNSCacheTime = map[string]time.Time{}
+
+var DNSCachemtx sync.RWMutex
+
+func handleDnsEvent(topic string, value string) {
+	var jsonData DNSRecord
+	if err := json.Unmarshal([]byte(value), &jsonData); err != nil {
+		log.Println("dns event, invalid json", err)
+		return
+	}
+
+	domain := ""
+	ips := []string{}
+	ttls := []int{}
+
+	for _, q := range jsonData.Q {
+		if q.Name == "" {
+			continue
+		}
+		domain = q.Name
+		if q.Qclass != 1 && q.Qclass != 28 {
+			continue
+		}
+
+		for _, a := range jsonData.A {
+			if a.Hdr.Class != 1 && a.Hdr.Class != 28 {
+				continue
+			}
+			if q.Name != "" && a.A != "" {
+				ip := net.ParseIP(a.A)
+				if ip != nil {
+					if ip.To4() != nil {
+						ips = append(ips, ip.To4().String())
+						ttls = append(ttls, a.Hdr.Ttl)
+					}
+				}
+			}
+		}
+
+		//only handle first query
+		break
+	}
+
+	if len(ips) != 0 {
+		if len(domain) > 1 {
+			//get rid if the pesky trailing .
+			domain = domain[:len(domain)-1]
+			updateDnsCache(domain, ips, ttls)
+		}
+	}
+}
+
+func updateDnsCache(domain string, ips []string, ttls []int) {
+	DNSCachemtx.Lock()
+	defer DNSCachemtx.Unlock()
+
+	for _, ip := range ips {
+		DNSCache[ip] = domain
+		DNSCacheTime[ip] = time.Now().Add(time.Duration(5) * time.Minute)
+	}
+}
+
+func cleanDNSCache() {
+	for {
+		time.Sleep(10 * time.Minute)
+		DNSCachemtx.Lock()
+		for key, value := range DNSCacheTime {
+			if time.Now().After(value) {
+				delete(DNSCacheTime, key)
+				delete(DNSCache, key)
+			}
+		}
+		DNSCachemtx.Unlock()
+	}
+}
+
+func DNSEventListener() {
+	go cleanDNSCache()
+	for i := 30; i > 0; i-- {
+		err := sprbus.HandleEvent("dns:serve:", handleDnsEvent)
+		if err != nil {
+			log.Println(err)
+		}
+		time.Sleep(3 * time.Second)
+	}
+	log.Println("[-] failed to establish connection to sprbus for dns")
 }