+++ frontend/src/components/Traffic/TimeSeriesList.js
@@ -16,11 +16,14 @@ import {
   Text,
   VStack
 } from '@gluestack-ui/themed'
-import { ModalContext } from 'AppContext'
+import { AppContext, ModalContext } from 'AppContext'
 import { copy } from 'utils'
+import DeviceItem from 'components/Devices/DeviceItem'
 
 const TimeSeriesList = ({ data, type, filterIps, setFilterIps, ...props }) => {
+  const context = useContext(AppContext)
   const modalContext = useContext(ModalContext)
+
   const regexLAN = /^192\.168\./ //TODO dont rely on this
 
   const [list, setList] = useState([])
@@ -95,13 +98,7 @@ const TimeSeriesList = ({ data, type, filterIps, setFilterIps, ...props }) => {
     return <></>
   })
 
-  const onPressIp = (e) => {
-    let ip = e.target.innerText
-
-    if (!ip) {
-      return
-    }
-
+  const onPressIp = (ip) => {
     if (ip.match(regexLAN) && setFilterIps) {
       setFilterIps([ip])
       return
@@ -131,40 +128,60 @@ const TimeSeriesList = ({ data, type, filterIps, setFilterIps, ...props }) => {
           justifyContent="space-between"
         >
           <VStack
-            sx={{ '@md': { flexDirection: 'row' } }}
-            flex={2}
+            flex={5}
             space="md"
             justifyContent="space-between"
+            sx={{ '@md': { flexDirection: 'row' } }}
           >
-            <HStack
-              flex={1}
+            <VStack
+              flex={2}
               space="md"
               justifyContent="space-between"
-              alignItems="center"
+              sx={{ '@md': { flexDirection: 'row', alignItems: 'center' } }}
             >
-              {/*['WanOut', 'LanOut', 'LanIn'].includes(type) ? (
-                  <Text bold>{item.deviceSrc && item.deviceSrc.Name}</Text>
-                ) : null*/}
-              <Pressable flex={1} onPress={onPressIp}>
-                <Text size="sm">{item.Src}</Text>
-              </Pressable>
-              <Box alignText="center">
+              <VStack sx={{ '@md': { flex: type.match(/Out$/) ? 2 : 3 } }}>
+                <Text size="sm" bold>
+                  {item.SrcDomain}
+                </Text>
+                {type == 'WanOut' ? (
+                  <DeviceItem
+                    item={context.getDevice(item.Src, 'RecentIP')}
+                    size="sm"
+                  />
+                ) : (
+                  <Text size="sm" onPress={() => onPressIp(item.Src)}>
+                    {item.Src}
+                  </Text>
+                )}
+              </VStack>
+              <Box
+                alignText="center"
+                display="none"
+                sx={{ '@md': { display: 'flex' } }}
+              >
                 <ArrowRightIcon color="$muted200" />
               </Box>
-              {/*['WanIn', 'LanOut', 'LanIn'].includes(type) ? (
-                  <Text bold>{item.deviceDst && item.deviceDst.Name}</Text>
-                ) : null*/}
-              <Pressable flex={1} onPress={onPressIp}>
-                <Text size="sm" textAlign="right">
-                  {item.Dst}
+              <VStack flex={type.match(/Out$/) ? 3 : 2}>
+                <Text size="sm" bold>
+                  {item.DstDomain}
                 </Text>
-              </Pressable>
-            </HStack>
+                {type == 'WanIn' ? (
+                  <DeviceItem
+                    item={context.getDevice(item.Dst, 'RecentIP')}
+                    size="sm"
+                  />
+                ) : (
+                  <Text size="sm" onPress={() => onPressIp(item.Dst)}>
+                    {item.Dst}
+                  </Text>
+                )}
+              </VStack>
+            </VStack>
             {showASN ? (
-              <HStack flex={1} alignItems="center">
+              <HStack flex={1} space="sm" alignItems="center">
                 {/*TODO also src depending on type*/}
                 <AsnIcon asn={item.Asn} />
-                <Text color="$muted500" size="sm" isTruncated>
+                <Text color="$muted500" size="xs" isTruncated>
                   {item.Asn}
                 </Text>
               </HStack>
@@ -179,7 +196,12 @@ const TimeSeriesList = ({ data, type, filterIps, setFilterIps, ...props }) => {
             sx={{ '@md': { justifyContent: 'flex-end' } }}
           >
             <Text size="xs">{timeAgo(item.Timestamp)}</Text>
-            <Badge action="muted" variant="outline" alignSelf="center">
+            <Badge
+              size="sm"
+              action="muted"
+              variant="outline"
+              alignSelf="center"
+            >
               <BadgeText>{prettySize(item.Bytes)}</BadgeText>
             </Badge>
           </HStack>