+++ frontend/src/views/Devices/AddDevice.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { useNavigate } from 'react-router-dom'
 
-import { View } from '@gluestack-ui/themed'
+import { ScrollView } from '@gluestack-ui/themed'
 
 import useSwipe from 'components/useSwipe'
 import AddDevice from 'components/Devices/AddDevice'
@@ -16,17 +16,16 @@ const AddDeviceView = () => {
   })
 
   return (
-    <View
+    <ScrollView
       bg="$backgroundCardLight"
+      h="$full"
       sx={{
         _dark: { bg: '$backgroundCardDark' }
       }}
-      p="$4"
-      h={'100%'}
       {...swipeHandlers}
     >
       <AddDevice />
-    </View>
+    </ScrollView>
   )
 }