+++ installer/scripts/unmount.sh
@@ -14,7 +14,7 @@ IMG="./data/spr.img"
 LOOP=$(losetup -j $IMG | cut -d: -f1)
 
 # shrink back img
-ef2sck -f ${LOOP}p2
+e2fsck -f ${LOOP}p2
 resize2fs ${LOOP}p2 7G
 
 losetup -d $LOOP 2>/dev/null